modify:20091109

+++ trunk/demo-apps/hpt/src/main/java/cn/hpt/model/PurchaseRecord.java
@@ -12,48 +12,48 @@ import javax.persistence.Table;
 @Table(name = "ht_purchase_record")
 public class PurchaseRecord {// 物资-详单表
 
-	@Id()
-	@SequenceGenerator(name = "ht_purchase_record_seq", sequenceName = "ht_purchase_record_seq")
-	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "ht_purchase_record_seq")
-	private Integer prnumber;
-	@ManyToOne
-	private Medicine medicine;
+    @Id()
+    @SequenceGenerator(name = "ht_purchase_record_seq", sequenceName = "ht_purchase_record_seq")
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "ht_purchase_record_seq")
+    private Long prnumber;
+    @ManyToOne
+    private Medicine medicine;
 
-	@ManyToOne
-	private Purchase purchase;
+    @ManyToOne
+    private Purchase purchase;
 
-	private Long prid;
+    private Long prid;
 
-	public Medicine getMedicine() {
-		return medicine;
-	}
+    public Medicine getMedicine() {
+        return medicine;
+    }
 
-	public void setMedicine(Medicine medicine) {
-		this.medicine = medicine;
-	}
+    public void setMedicine(Medicine medicine) {
+        this.medicine = medicine;
+    }
 
-	public Integer getPrnumber() {
-		return prnumber;
-	}
+    public Long getPrnumber() {
+        return prnumber;
+    }
 
-	public void setPrnumber(Integer prnumber) {
-		this.prnumber = prnumber;
-	}
+    public void setPrnumber(Long prnumber) {
+        this.prnumber = prnumber;
+    }
 
-	public Long getPrid() {
-		return prid;
-	}
+    public Long getPrid() {
+        return prid;
+    }
 
-	public void setPrid(Long prid) {
-		this.prid = prid;
-	}
+    public void setPrid(Long prid) {
+        this.prid = prid;
+    }
 
-	public Purchase getPurchase() {
-		return purchase;
-	}
+    public Purchase getPurchase() {
+        return purchase;
+    }
 
-	public void setPurchase(Purchase purchase) {
-		this.purchase = purchase;
-	}
+    public void setPurchase(Purchase purchase) {
+        this.purchase = purchase;
+    }
 
 }
